Limit recursion depth of file history to 20 files
When project (baseline) history is broken, the file recursion might recurse the entire tree og the file history.
This is probably not wanted functionallity, so the recursion limit is set at a depths of 20 predecessors for
each file changed in the project.
If the recursion depth is exceeded only the version of the file in the current project is returned.